move async thunk generation code into separate file
